% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_kernel_pls}
\alias{cpp_kernel_pls}
\title{Internal kernel and wide-kernel PLS solver}
\usage{
cpp_kernel_pls(X, Y, ncomp, tol, wide)
}
\arguments{
\item{X}{Centered design matrix.}

\item{Y}{Centered response matrix.}

\item{ncomp}{Maximum number of components.}

\item{tol}{Numerical tolerance.}

\item{wide}{Whether to use the wide-kernel update.}
}
\value{
A list containing the kernel PLS factors.
}
\description{
Internal kernel and wide-kernel PLS solver
}
