% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_fit.R
\name{.finalize_pls_fit}
\alias{.finalize_pls_fit}
\title{Finalize pls objects}
\usage{
.finalize_pls_fit(fit, algorithm)
}
\arguments{
\item{fit}{Fitted object}

\item{algorithm}{Name of the algorithm used to fit the object}
}
\value{
The \code{fit} object with normalized naming and class attributes.
}
\description{
Finalize pls objects
}
