% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{.resolve_training_ref}
\alias{.resolve_training_ref}
\title{Internal: resolve training reference for RKHS predictions}
\usage{
.resolve_training_ref(obj, dots)
}
\description{
Accepts:
\itemize{
\item dense matrix (returned as-is)
\item big.matrix (returned as-is)
\item big.matrix.descriptor (attached and returned)
}
}
\details{
Sources (priority):
object$X, object$Xtrain, ...$Xtrain, ...$X_ref, object$X_ref
}
\keyword{internal}
