% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_analysis.R
\name{plot_pls_individuals}
\alias{plot_pls_individuals}
\title{Plot individual scores}
\usage{
plot_pls_individuals(object, comps = c(1L, 2L), labels = NULL, ...)
}
\arguments{
\item{object}{A fitted PLS model with scores.}

\item{comps}{Components to plot (length two).}

\item{labels}{Optional character vector of point labels.}

\item{...}{Additional plotting parameters passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\description{
Plot individual scores
}
