% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsR2.R
\name{pls2_dense}
\alias{pls2_dense}
\alias{pls2_stream}
\title{Partial least squares regression for multi-response problems (PLS2)}
\usage{
pls2_dense(
  X,
  Y,
  ncomp,
  center = TRUE,
  scale = FALSE,
  algorithm = c("simpls", "nipals"),
  return_big = FALSE
)

pls2_stream(
  X,
  Y,
  ncomp,
  center = TRUE,
  scale = FALSE,
  block_size = 1024L,
  algorithm = c("simpls", "nipals"),
  return_big = FALSE
)
}
\arguments{
\item{X}{A \code{bigmemory::big.matrix} containing the predictor variables.}

\item{Y}{A \code{bigmemory::big.matrix} storing the multi-dimensional response.}

\item{ncomp}{Number of latent components to compute.}

\item{center}{Should the inputs be centered prior to fitting?}

\item{scale}{Should the inputs be scaled to unit variance prior to fitting?}

\item{algorithm}{PLS backend to use. Either "simpls" (default) or "nipals".}

\item{return_big}{Logical; when \code{TRUE}, the coefficients, scores and loadings
are returned as \code{\link[bigmemory:big.matrix]{bigmemory::big.matrix}} objects. Defaults to \code{FALSE}.}

\item{block_size}{Number of rows processed per block by the streaming
variant.}
}
\value{
A list with regression coefficients, intercept, weights, loadings
and preprocessing metadata.
}
\description{
Partial least squares regression for multi-response problems (PLS2)
}
