% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_analysis.R
\name{pls_predict_response}
\alias{pls_predict_response}
\title{Predict responses from a PLS fit}
\usage{
pls_predict_response(object, newdata, ncomp = NULL)
}
\arguments{
\item{object}{A fitted PLS model.}

\item{newdata}{Predictor matrix for scoring.}

\item{ncomp}{Number of components to use.}
}
\value{
A numeric matrix or vector of predictions.
}
\description{
Predict responses from a PLS fit
}
