% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_analysis.R
\name{pls_threshold}
\alias{pls_threshold}
\title{Naive sparsity control by coefficient thresholding}
\usage{
pls_threshold(object, threshold)
}
\arguments{
\item{object}{A fitted PLS model.}

\item{threshold}{Values below this absolute magnitude are set to zero.}
}
\value{
A modified copy of \code{object} with thresholded coefficients.
}
\description{
Naive sparsity control by coefficient thresholding
}
