% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigPLScpp-package.R
\docType{package}
\name{bigPLSR-package}
\alias{bigPLSR-package}
\alias{bigPLSR}
\title{bigPLSR-package}
\description{
Provides Partial least squares Regression for big data. It allows for missing data in the explanatory variables. Repeated k-fold cross-validation of such models using various criteria. Bootstrap confidence intervals constructions are also available.
}
\examples{
set.seed(314)
library(bigPLSR)
data(sim_data)
head(sim_data)

}
\references{
Maumy, M., Bertrand, F. (2023). PLS models and their extension for big data.
Joint Statistical Meetings (JSM 2023), Toronto, ON, Canada.

Maumy, M., Bertrand, F. (2023). bigPLS: Fitting and cross-validating
PLS-based Cox models to censored big data. BioC2023 — The Bioconductor
Annual Conference, Dana-Farber Cancer Institute, Boston, MA, USA.
Poster. https://doi.org/10.7490/f1000research.1119546.1
}
\seealso{
\code{\link[=big_plsR]{big_plsR()}} and \code{\link[=big_plsR_gd]{big_plsR_gd()}}
}
\author{
\strong{Maintainer}: Frederic Bertrand \email{frederic.bertrand@lecnam.net} (\href{https://orcid.org/0000-0002-0837-8281}{ORCID})

Authors:
\itemize{
  \item Myriam Maumy \email{myriam.maumy@ehesp.fr} (\href{https://orcid.org/0000-0002-4615-1512}{ORCID})
}

}
