% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_analysis.R
\name{plot_pls_biplot}
\alias{plot_pls_biplot}
\title{PLS biplot}
\usage{
plot_pls_biplot(object, comps = c(1L, 2L), scale_variables = 1, ...)
}
\arguments{
\item{object}{A fitted PLS model with scores and loadings.}

\item{comps}{Components to display.}

\item{scale_variables}{Scaling factor applied to variable loadings.}

\item{...}{Additional arguments passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\description{
PLS biplot
}
