% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_analysis.R
\name{plot_pls_variables}
\alias{plot_pls_variables}
\title{Plot variable loadings}
\usage{
plot_pls_variables(object, comps = c(1L, 2L), ...)
}
\arguments{
\item{object}{A fitted PLS model.}

\item{comps}{Components to display (length two).}

\item{...}{Additional plotting parameters passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\description{
Plot variable loadings
}
