% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_analysis.R
\name{pls_predict_scores}
\alias{pls_predict_scores}
\title{Predict latent scores from a PLS fit}
\usage{
pls_predict_scores(object, newdata, ncomp = NULL)
}
\arguments{
\item{object}{A fitted PLS model.}

\item{newdata}{Predictor matrix for scoring.}

\item{ncomp}{Number of components to use.}
}
\value{
Matrix of component scores.
}
\description{
Predict latent scores from a PLS fit
}
