% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_analysis.R
\name{summary.big_plsr}
\alias{summary.big_plsr}
\title{Summaries for big_plsr objects}
\usage{
\method{summary}{big_plsr}(object, X = NULL, Y = NULL, ...)
}
\arguments{
\item{object}{A fitted PLS model.}

\item{X}{Optional design matrix to recompute reconstruction metrics.}

\item{Y}{Optional response matrix/vector.}

\item{...}{Unused.}
}
\value{
An object of class \code{summary.big_plsr}.
}
\description{
Summaries for big_plsr objects
}
